not yet finished with add add new request result in Json file

+++ Generation_2_lead/example/FileManager.py
@@ -7,13 +7,20 @@ class FileManager():
 
     def WriteInFile(self,data,enterUrl,LastpageNbr):
         os.chdir(r'C:\Users\Ranyl\Desktop\tets')
-
-        try:
-            with open("{}.json".format(enterUrl), 'w') as outfile:
+        if self.verifyIfFileExist(self,enterUrl):
+            with open("{}.json".format(enterUrl), 'a+') as outfile:
+                outfile.seek(0, 2)
+                outfile.truncate()
                 data.append({"LastpageNbr": LastpageNbr})
                 json.dump(data, outfile)
-        except FileNotFoundError:
-            pass
+
+        else:
+            try:
+                with open("{}.json".format(enterUrl), 'w') as outfile:
+                    data.append({"LastpageNbr": LastpageNbr})
+                    json.dump(data, outfile)
+            except FileNotFoundError:
+                pass
 
 
     def GetLastPageNumber(self, enterUrl):
